Int -> Double for antall.

+++ src/main/kotlin/no/nav/tms/utbetalingsoversikt/api/ytelse/domain/external/YtelseskomponentEkstern.kt
@@ -7,6 +7,6 @@ data class YtelseskomponentEkstern(
     val ytelseskomponenttype: String? = null,
     val satsbeloep: Double? = null,
     val satstype: String? = null,
-    val satsantall: Int? = null,
+    val satsantall: Double? = null,
     val ytelseskomponentbeloep: Double? = null,
 )
